+++ src/test/java/ispd/application/terminal/ModelFolder.java
@@ -1,0 +1,17 @@
+package ispd.application.terminal;
+
+import java.nio.file.*;
+
+public enum ModelFolder {
+    NO_TYPE("notype");
+
+    private final Path path;
+
+    ModelFolder (final String folderName) {
+        this.path = TerminalApplicationCharacterizationTest.MODEL_FOLDER_PATH.resolve(folderName);
+    }
+
+    public String pathTo (final String modelName) {
+        return this.path.resolve(modelName).toString();
+    }
+}
